Move indexation services of events to separate service provider
